Add files via upload
Restaurant reviews and menu items were matched to find out the ratings of the menu items from the reviews. Then average ratings and number of reviews per item were calculated. This gave us the items with low reviews so that the client can than decide which menu items need to be modified or deleted.